+++ packages/components/src/components/Action/types.ts
@@ -3,16 +3,20 @@ import type { OverlayController } from "@/lib/controller";
 import type { ActionModel } from "@/components/Action/models/ActionModel";
 import type { FlowComponentName } from "@/components/propTypes";
 import type { FlowComponentProps } from "@/lib/componentFactory/flowComponent";
+import type {
+  CloseModalOptions,
+  CloseOverlayOptions,
+} from "@/lib/controller/overlay/OverlayController";
 
 export type ActionFn = (...args: unknown[]) => unknown;
 
 export interface ActionProps extends PropsWithChildren, FlowComponentProps {
   onAction?: ActionFn;
   actionModel?: ActionModel;
-  closeOverlay?: FlowComponentName | OverlayController;
+  closeOverlay?: FlowComponentName | OverlayController | CloseOverlayOptions;
   openOverlay?: FlowComponentName | OverlayController;
   toggleOverlay?: FlowComponentName | OverlayController;
-  closeModal?: boolean;
+  closeModal?: boolean | CloseModalOptions;
   openModal?: boolean;
   toggleModal?: boolean;
   break?: boolean;