+++ packages/components/src/integrations/react-hook-form/index.ts
@@ -6,3 +6,5 @@ export {
   useFormContext,
   useOptionalFormContext,
 } from "./components/FormContextProvider";
+export * from "./components/FormAction";
+export { flags } from "./flags";