+++ packages/components/src/lib/controller/overlay/OverlayController.ts
@@ -2,6 +2,8 @@ import { action, makeObservable, observable } from "mobx";
 import useSelector from "@/lib/mobx/useSelector";
 import { useStatic } from "@/lib/hooks/useStatic";
 import { useEffect } from "react";
+import { useCloseOverlayConfirmationController } from "@/lib/controller/overlay/useCloseOverlayConfirmationController";
+import type { FlowComponentName } from "@/components/propTypes";
 
 export type OverlayOpenHandler = () => unknown;
 export type OverlayCloseHandler = () => unknown;
@@ -13,44 +15,76 @@ type AnyOverlayOpenStateHandler =
 
 type DisposerFn = () => void;
 
+export interface CloseOverlayOptions {
+  overlay: FlowComponentName | OverlayController;
+  bypassConfirmation?: boolean;
+}
+
+export type CloseModalOptions = Omit<CloseOverlayOptions, "overlay">;
+
+type CloseOptions = CloseOverlayOptions | CloseModalOptions;
+
 export interface OverlayControllerOptions {
   isDefaultOpen?: boolean;
   onOpen?: OverlayOpenHandler;
   onClose?: OverlayCloseHandler;
   onOpenChange?: OverlayOpenStateHandler;
+  confirmOnClose?: boolean;
 }
+
+type ConstructorOptions = Pick<
+  OverlayControllerOptions,
+  "isDefaultOpen" | "confirmOnClose"
+>;
 
 export class OverlayController {
   public isOpen = false;
+  public showConfirmationModal = false;
+  public confirmOnCloseEnabled: boolean;
   private onOpenHandlers = new Set<OverlayOpenHandler>();
   private onCloseHandlers = new Set<OverlayCloseHandler>();
   private onOpenChangeHandlers = new Set<OverlayOpenStateHandler>();
 
-  public constructor(options?: OverlayControllerOptions) {
+  public constructor(options: ConstructorOptions = {}) {
     makeObservable(this, {
       isOpen: observable,
+      showConfirmationModal: observable,
       open: action.bound,
       close: action.bound,
       toggle: action.bound,
       setOpen: action.bound,
+      confirmClose: action.bound,
     });
     this.isOpen = options?.isDefaultOpen ?? false;
+    this.confirmOnCloseEnabled = options?.confirmOnClose === true;
+  }
 
-    if (options?.onOpen) {
-      this.onOpenHandlers.add(options.onOpen);
-    }
-    if (options?.onClose) {
-      this.onCloseHandlers.add(options.onClose);
-    }
-    if (options?.onOpenChange) {
-      this.onOpenChangeHandlers.add(options.onOpenChange);
-    }
+  public useUpdateOptions(options: OverlayControllerOptions = {}): void {
+    const { onOpen, onClose, onOpenChange, confirmOnClose } = options;
+
+    this.useOnHandler(onOpen, (h) =>
+      this.addOpenStateHandler(h, this.onOpenHandlers),
+    );
+    this.useOnHandler(onClose, (h) =>
+      this.addOpenStateHandler(h, this.onCloseHandlers),
+    );
+    this.useOnHandler(onOpenChange, (h) =>
+      this.addOpenStateHandler(h, this.onOpenChangeHandlers),
+    );
+
+    useEffect(() => {
+      if (confirmOnClose) {
+        this.confirmOnCloseEnabled = true;
+      }
+    }, [confirmOnClose, this]);
   }
 
   public static useNew(
     options: OverlayControllerOptions = {},
   ): OverlayController {
-    return useStatic(() => new OverlayController(options));
+    const controller = useStatic(() => new OverlayController(options));
+    controller.useUpdateOptions(options);
+    return controller;
   }
 
   private addOpenStateHandler<T extends AnyOverlayOpenStateHandler>(
@@ -63,18 +97,6 @@ export class OverlayController {
     };
   }
 
-  public addOnOpen(handler: OverlayOpenHandler): DisposerFn {
-    return this.addOpenStateHandler(handler, this.onOpenHandlers);
-  }
-
-  public addOnClose(handler: OverlayCloseHandler): DisposerFn {
-    return this.addOpenStateHandler(handler, this.onCloseHandlers);
-  }
-
-  public addOnOpenChange(handler: OverlayOpenStateHandler): DisposerFn {
-    return this.addOpenStateHandler(handler, this.onOpenChangeHandlers);
-  }
-
   private useOnHandler<T extends AnyOverlayOpenStateHandler>(
     handler: T | undefined,
     addHandlerFn: (handler: T) => DisposerFn,
@@ -83,18 +105,6 @@ export class OverlayController {
       () => (handler ? addHandlerFn(handler) : undefined),
       [handler, this],
     );
-  }
-
-  public useOnOpen(handler?: OverlayOpenHandler): void {
-    this.useOnHandler(handler, (h) => this.addOnOpen(h));
-  }
-
-  public useOnClose(handler?: OverlayCloseHandler): void {
-    this.useOnHandler(handler, (h) => this.addOnClose(h));
-  }
-
-  public useOnOpenChange(handler?: OverlayOpenStateHandler): void {
-    this.useOnHandler(handler, (h) => this.addOnOpenChange(h));
   }
 
   private executeHandlers(
@@ -121,39 +131,79 @@ export class OverlayController {
     return this.executeHandlers(isOpen, this.onOpenChangeHandlers);
   }
 
+  public addOnClose(handler: OverlayCloseHandler) {
+    return this.addOpenStateHandler(handler, this.onCloseHandlers);
+  }
+
+  public addOnOpen(handler: OverlayOpenHandler) {
+    return this.addOpenStateHandler(handler, this.onOpenHandlers);
+  }
+
+  public addOnOpenChange(handler: OverlayOpenStateHandler) {
+    return this.addOpenStateHandler(handler, this.onOpenChangeHandlers);
+  }
+
   public open(): void {
     this.setOpen(true);
   }
 
-  public close(): void {
-    this.setOpen(false);
+  public close(options?: CloseOptions): void {
+    this.setOpen(false, options);
   }
 
   public toggle(): void {
     this.setOpen(!this.isOpen);
   }
 
-  public setOpen(to: boolean): void {
-    if (this.isOpen === to) {
+  public setOpen(toOpen: boolean, options: CloseOptions = {}): void {
+    if (this.isOpen === toOpen) {
+      return;
+    }
+
+    const { bypassConfirmation = false } = options;
+
+    if (
+      this.confirmOnCloseEnabled &&
+      toOpen === false &&
+      bypassConfirmation === false
+    ) {
+      this.showConfirmationModal = true;
       return;
     }
 
     let aborted = false;
-    if (to) {
+    if (toOpen) {
       aborted = this.executeOnOpen();
     } else {
       aborted = this.executeOnClose();
     }
     if (!aborted) {
-      aborted = this.executeOnOpenChange(to);
+      aborted = this.executeOnOpenChange(toOpen);
     }
 
     if (!aborted) {
-      this.isOpen = to;
+      this.isOpen = toOpen;
     }
   }
 
   public useIsOpen() {
     return useSelector(() => this.isOpen);
+  }
+
+  public useShowConfirmationModal() {
+    return useSelector(() => this.showConfirmationModal);
+  }
+
+  public useConfirmationController() {
+    return useCloseOverlayConfirmationController(this);
+  }
+
+  public confirmClose(): void {
+    this.showConfirmationModal = false;
+    this.confirmOnCloseEnabled = false;
+  }
+
+  public cancelConfirmation(): void {
+    this.showConfirmationModal = false;
   }
 }